refactor: déplace enum email
