+++ backend/lib/email.ts
@@ -1,6 +1,6 @@
 import { ArgumentParser } from "argparse"
 import config from "../config/index.js"
-import { EmailCategory } from "../enums/email.js"
+import { EmailCategory } from "../../lib/enums/email.js"
 import mongoose from "mongoose"
 import mongooseConfig from "../config/mongoose.js"
 import { processSendEmails } from "./emails/sending.js"
