+++ backend/models/followup.ts
@@ -6,7 +6,7 @@ import { Survey } from "../../lib/types/survey.js"
 import { SurveyCategory } from "../../lib/enums/survey.js"
 import emailRender from "../lib/mes-aides/emails/email-render.js"
 import SurveySchema from "./survey-schema.js"
-import { EmailCategory } from "../enums/email.js"
+import { EmailCategory } from "../../lib/enums/email.js"
 import config from "../config/index.js"
 import { Followup } from "../../lib/types/followup.d.js"
 import { FollowupModel } from "../types/models.d.js"
