+++ lib/enums/email.ts
